SVG Export
I noticed that the exportSVG function wraps the X and Y axis titles in <span></span> tags.  This cause the exported axis titles, graph titles, and graphing increments to appear jumbled and centered beneath the actual graph.  Removing the addition of span tags in the JS on line 2319 fixes this issue and allows the svg graph to be formatted correctly